requirements.txt
python-dotenv==1.2.2
requests==2.32.5
redis==5.2.1
